Group the primitives into a query_metrics package and rename for clarity
The two modules were named for their mechanisms rather than their subject, which
left the vocabulary inconsistent with the rest of DBM and made the classification
enum read as an instruction to the base library rather than a description of what
the integration saw.

- db/query_metrics/ now holds the set, split into stats, obfuscation, cache and
  resolver. The package re-exports the public names lazily, mirroring db/__init__.py,
  so the file layout is not part of the interface.
- DeltaDetector.compute() becomes QueryStats.diff(snapshot), returning a Delta.
  "Query stats" is what every source calls itself (pg_stat_statements,
  dm_exec_query_stats, $queryStats), and it keeps stats (what the database exposes)
  distinct from metrics (what we emit). derivative_rows becomes rows, and
  metric_columns becomes counter_columns to match.
- TextDisposition becomes TextKind, with STATEMENT/EXCLUDED/UNAVAILABLE replacing
  CACHE/IGNORE/SKIP. Integrations now report what a text turned out to be and the
  resolver owns the caching policy those kinds imply, rather than each integration
  re-deriving whether a rejection is permanent. The kinds split on whether the text
  is the statement's own and, if not, whether that can change; text that is
  permanently unavailable would need a kind of its own.
