+++ datadog_checks_base/datadog_checks/base/utils/db/query_metrics/__init__.pyi
@@ -1,0 +1,19 @@
+# (C) Datadog, Inc. 2026-present
+# All rights reserved
+# Licensed under a 3-clause BSD style license (see LICENSE)
+from .cache import ObfuscationLookup
+from .obfuscation import ObfuscationResult, obfuscate_statement
+from .resolver import ResolveResult, ResolveStats, TextKind, resolve_obfuscations
+from .stats import Delta, QueryStats
+
+__all__ = [
+    'Delta',
+    'ObfuscationLookup',
+    'ObfuscationResult',
+    'QueryStats',
+    'ResolveResult',
+    'ResolveStats',
+    'TextKind',
+    'obfuscate_statement',
+    'resolve_obfuscations',
+]
