+++ datadog_checks_base/datadog_checks/base/utils/db/query_metrics/stats.py
@@ -11,15 +11,16 @@ logger = logging.getLogger(__name__)
 
 
 @dataclass
-class DeltaResult[K: Hashable]:
-    derivative_rows: list[dict]
-    """Rows whose counters advanced since the previous snapshot, with metric columns replaced by their deltas."""
+class Delta[K: Hashable]:
+    rows: list[dict]
+    """Rows whose counters advanced since the previous snapshot, with counter columns replaced by
+    their deltas."""
     changed_keys: set[K]
-    """Keys of the rows in ``derivative_rows``."""
+    """Keys of :attr:`rows`, for resolving statement text only for what actually ran."""
 
 
-class DeltaDetector[K: Hashable]:
-    """Diffs consecutive snapshots of cumulative counters to produce per-key derivative rows.
+class QueryStats[K: Hashable]:
+    """Diffs consecutive snapshots of cumulative query statistics into per-interval rows.
 
     Statistics tables such as ``pg_stat_statements``, MySQL's
     ``events_statements_summary_by_digest`` and SQL Server's ``dm_exec_query_stats`` expose
@@ -35,11 +36,11 @@ class DeltaDetector[K: Hashable]:
 
     def __init__(
         self,
-        metric_columns: frozenset[str] | set[str],
+        counter_columns: frozenset[str] | set[str],
         key: Callable[[dict], K],
         execution_indicators: frozenset[str] | set[str] | None = None,
     ):
-        self._metric_columns = frozenset(metric_columns)
+        self._counter_columns = frozenset(counter_columns)
         self._key = key
         # Counters that only advance when a statement executes. When set, a key whose indicators are
         # all flat is skipped without diffing the remaining columns.
@@ -49,14 +50,14 @@ class DeltaDetector[K: Hashable]:
     def reset(self):
         self._previous.clear()
 
-    def compute(self, rows: list[dict]) -> DeltaResult[K]:
-        """Diff *rows* against the previous snapshot and remember them for the next call."""
-        current = self._collapse(rows)
+    def diff(self, snapshot: list[dict]) -> Delta[K]:
+        """Diff *snapshot* against the previous one and remember it for the next call."""
+        current = self._collapse(snapshot)
 
-        derivative_rows: list[dict] = []
+        rows: list[dict] = []
         changed_keys: set[K] = set()
 
-        available_metrics: frozenset[str] | None = None
+        available_counters: frozenset[str] | None = None
         indicator_cols: frozenset[str] | None = None
 
         for key, row in current.items():
@@ -65,16 +66,16 @@ class DeltaDetector[K: Hashable]:
                 # First time we've seen this key; there is no baseline to subtract yet.
                 continue
 
-            if available_metrics is None:
-                available_metrics = self._metric_columns & row.keys() & prev.keys()
-                indicator_cols = self._execution_indicators & available_metrics
+            if available_counters is None:
+                available_counters = self._counter_columns & row.keys() & prev.keys()
+                indicator_cols = self._execution_indicators & available_counters
 
             if indicator_cols and not any(row[col] - prev[col] > 0 for col in indicator_cols):
                 continue
 
             has_negative = False
             has_change = False
-            for col in available_metrics:
+            for col in available_counters:
                 diff = row[col] - prev[col]
                 if diff < 0:
                     # A counter went backwards, so the series was reset and the diff is meaningless.
@@ -87,46 +88,46 @@ class DeltaDetector[K: Hashable]:
             if has_negative or not has_change:
                 continue
 
-            derivative = {}
+            delta_row = {}
             for col in row:
-                if col in available_metrics:
-                    derivative[col] = row[col] - prev[col]
+                if col in available_counters:
+                    delta_row[col] = row[col] - prev[col]
                 else:
-                    derivative[col] = row[col]
-            derivative_rows.append(derivative)
+                    delta_row[col] = row[col]
+            rows.append(delta_row)
             changed_keys.add(key)
 
         logger.debug(
-            "delta: snapshot=%d prev=%d derivative=%d changed=%d",
+            "diff: snapshot=%d prev=%d rows=%d changed=%d",
             len(current),
             len(self._previous),
-            len(derivative_rows),
+            len(rows),
             len(changed_keys),
         )
 
-        self._update_cache(current)
+        self._remember(current)
 
-        return DeltaResult(derivative_rows=derivative_rows, changed_keys=changed_keys)
+        return Delta(rows=rows, changed_keys=changed_keys)
 
-    def _collapse(self, rows: list[dict]) -> dict[K, dict]:
-        """Group rows by key, summing metric columns across rows that share one.
+    def _collapse(self, snapshot: list[dict]) -> dict[K, dict]:
+        """Group rows by key, summing counter columns across rows that share one.
 
         Rows are copied rather than accumulated into, so a caller that holds on to the snapshot it
         passed in does not see its counters silently rewritten.
         """
         collapsed: dict[K, dict] = {}
-        for row in rows:
+        for row in snapshot:
             key = self._key(row)
             existing = collapsed.get(key)
             if existing is None:
                 collapsed[key] = dict(row)
                 continue
-            for col in self._metric_columns:
+            for col in self._counter_columns:
                 if col in row:
                     existing[col] = existing.get(col, 0) + row[col]
         return collapsed
 
-    def _update_cache(self, current: dict[K, dict]):
+    def _remember(self, current: dict[K, dict]):
         stale = self._previous.keys() - current.keys()
         for k in stale:
             del self._previous[k]
@@ -134,11 +135,11 @@ class DeltaDetector[K: Hashable]:
         for key, row in current.items():
             prev = self._previous.get(key)
             if prev is not None:
-                # Only the metric columns are retained; everything else is re-read each cycle.
-                for col in self._metric_columns:
+                # Only the counter columns are retained; everything else is re-read each cycle.
+                for col in self._counter_columns:
                     if col in row:
                         prev[col] = row[col]
                     elif col in prev:
                         del prev[col]
             else:
-                self._previous[key] = {col: row[col] for col in self._metric_columns if col in row}
+                self._previous[key] = {col: row[col] for col in self._counter_columns if col in row}
