+++ datadog_checks_base/tests/base/utils/db/query_metrics/test_cache.py
@@ -5,28 +5,7 @@
 
 from unittest import mock
 
-from datadog_checks.base.utils.db.obfuscation_lookup import ObfuscationLookup, obfuscate_statement
-
-
-class TestObfuscateStatement:
-    def test_returns_result_with_signature(self):
-        result = obfuscate_statement('SELECT 1', '{}')
-        assert result is not None
-        assert result.obfuscated_query
-        assert result.query_signature
-
-    def test_returns_none_when_obfuscation_fails(self):
-        with mock.patch(
-            'datadog_checks.base.utils.db.obfuscation_lookup.obfuscate_sql_with_metadata',
-            side_effect=RuntimeError('cannot obfuscate'),
-        ):
-            assert obfuscate_statement('SELECT 1', '{}') is None
-
-    def test_identical_text_yields_identical_signature(self):
-        """Callers that cannot cache still get signatures consistent with the cached path."""
-        first = obfuscate_statement('SELECT 1', '{}')
-        second = obfuscate_statement('SELECT 1', '{}')
-        assert first.query_signature == second.query_signature
+from datadog_checks.base.utils.db.query_metrics import ObfuscationLookup
 
 
 class TestObfuscationLookup:
@@ -112,7 +91,7 @@ class TestObfuscationLookup:
             return {'query': raw_text, 'metadata': {}}
 
         with mock.patch(
-            'datadog_checks.base.utils.db.obfuscation_lookup.obfuscate_sql_with_metadata',
+            'datadog_checks.base.utils.db.query_metrics.obfuscation.obfuscate_sql_with_metadata',
             side_effect=obfuscate,
         ):
             results, failures = lk.populate({(1, 1, 1): 'SELECT 1', (2, 1, 1): 'BAD'})
